backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.128.0",
    "psycopg2>=2.9.11",
    "pyjwt[crypto]>=2.10.1",
    "redis>=7.1.0",
    "sqlmodel>=0.0.31",
]
